feat(ui): compact drawer forms with single-line rows and collapsible sections
- Collapse production goal and input item rows to a single line using placeholders instead of per-field labels
- Move section descriptions into an info-icon tooltip beside each title via a reusable SectionTitle component
- Move Weighting Options and Transport Capacity sections to the Production tab below Production Goals
- Add CollapsibleSection wrapper so all drawer sections can collapse/expand
