+++ client/src/containers/ProductionPlanner/PlannerOptions/InputsTab/index.tsx
@@ -1,41 +1,9 @@
 import React, { useMemo } from 'react';
 import styled from 'styled-components';
-import { Button, Select, TextInput, Checkbox, Group, Title, Text } from '@mantine/core';
+import { Button, Select, TextInput, Checkbox, Group, Text } from '@mantine/core';
 import { useProductionContext } from '../../../../contexts/production';
 import TrashButton from '../../../../components/TrashButton';
-import { Section } from '../../../../components/Section';
-import LabelWithTooltip from '../../../../components/LabelWithTooltip';
-import { TransportOptions } from '../../../../contexts/production/types';
-
-const BELT_TIER_OPTIONS = [
-  { value: 'disabled', label: 'Disabled' },
-  { value: '60', label: 'Mk. 1 (60/min)' },
-  { value: '120', label: 'Mk. 2 (120/min)' },
-  { value: '240', label: 'Mk. 3 (240/min)' },
-  { value: '480', label: 'Mk. 4 (480/min)' },
-  { value: '780', label: 'Mk. 5 (780/min)' },
-  { value: '1200', label: 'Mk. 6 (1200/min)' },
-  { value: 'custom', label: 'Custom' },
-];
-
-const BELT_PRESET_VALUES = new Set(['60', '120', '240', '480', '780', '1200']);
-
-const PIPE_TIER_OPTIONS = [
-  { value: 'disabled', label: 'Disabled' },
-  { value: '300', label: 'Mk. 1 (300 m³/min)' },
-  { value: '600', label: 'Mk. 2 (600 m³/min)' },
-  { value: 'custom', label: 'Custom' },
-];
-
-const PIPE_PRESET_VALUES = new Set(['300', '600']);
-
-function getSelectValue(capacity: string | null, presets: Set<string>): string {
-  if (capacity === null) return 'disabled';
-  if (presets.has(capacity)) return capacity;
-  return 'custom';
-}
-
-
+import { CollapsibleSection } from '../../../../components/Section';
 
 const InputsTab = () => {
   const ctx = useProductionContext();
@@ -54,10 +22,9 @@ const InputsTab = () => {
     return ctx.state.inputItems.map((data) => {
       return (
         <ItemContainer key={data.key}>
-          <Row>
+          <Row wrap='nowrap' align='center' gap='xs'>
             <Select
               placeholder="Select an item"
-              label='Item'
               clearable
               searchable
               data={itemOptions}
@@ -68,13 +35,10 @@ const InputsTab = () => {
                   data: { ...data, itemKey: value || '' },
                 });
               }}
-              style={{ flex: '1 1 auto' }}
+              style={{ flex: '1 1 0' }}
             />
-            <TrashButton onClick={() => { ctx.dispatch({ type: 'DELETE_INPUT_ITEM', key: data.key }); }} style={{ position: 'relative', top: '13px' }} />
-          </Row>
-          <Row>
             <TextInput
-              label='Amount'
+              placeholder='Amount'
               className='no-spinner'
               type='number'
               min='0'
@@ -87,6 +51,7 @@ const InputsTab = () => {
                   data: { ...data, value: e.currentTarget.value },
                 });
               }}
+              style={{ width: '110px' }}
             />
             <Checkbox
               className='label'
@@ -98,156 +63,12 @@ const InputsTab = () => {
                   data: { ...data, unlimited: !data.unlimited },
                 });
               }}
-              style={{ position: 'relative', top: '13px' }}
             />
+            <TrashButton onClick={() => { ctx.dispatch({ type: 'DELETE_INPUT_ITEM', key: data.key }); }} />
           </Row>
         </ItemContainer>
       )
     });
-  }
-
-  function renderWeightInputs() {
-    const weightingOptions = ctx.state.weightingOptions;
-    return (
-      <>
-        <Group grow>
-          <TextInput
-            label={<LabelWithTooltip label='Resource Efficiency' tooltip='This weighting prioritizes using as few resources as possible.' />}
-            className='no-spinner'
-            type='number'
-            min='0'
-            step='1'
-            value={weightingOptions.resources}
-            onChange={(e) => {
-              ctx.dispatch({
-                type: 'UPDATE_WEIGHTING_OPTIONS',
-                data: { ...weightingOptions, resources: e.currentTarget.value },
-              });
-            }}
-          />
-          <TextInput
-            label={<LabelWithTooltip label='Power Efficiency' tooltip='This weighting prioritizes using as little power as possible.' />}
-            className='no-spinner'
-            type='number'
-            min='0'
-            step='1'
-            value={weightingOptions.power}
-            onChange={(e) => {
-              ctx.dispatch({
-                type: 'UPDATE_WEIGHTING_OPTIONS',
-                data: { ...weightingOptions, power: e.currentTarget.value },
-              });
-            }}
-          />
-        </Group>
-        <Group grow style={{ marginTop: '10px' }}>
-          <TextInput
-            label={<LabelWithTooltip label='Complexity' tooltip='This weighting prioritizes reducing the number of item types used in the factory. Very slow to optimize for larger factories (WIP).' />}
-            className='no-spinner'
-            type='number'
-            min='0'
-            step='1'
-            value={weightingOptions.complexity}
-            onChange={(e) => {
-              ctx.dispatch({
-                type: 'UPDATE_WEIGHTING_OPTIONS',
-                data: { ...weightingOptions, complexity: e.currentTarget.value },
-              });
-            }}
-          />
-          <TextInput
-            label={<LabelWithTooltip label='Buildings' tooltip='This weighting prioritizes using as few buildings as possible, discounting overclocking. May not be perfectly optimal, especially for smaller factories (WIP).' />}
-            type='number'
-            min='0'
-            step='1'
-            value={weightingOptions.buildings}
-            onChange={(e) => {
-              ctx.dispatch({
-                type: 'UPDATE_WEIGHTING_OPTIONS',
-                data: { ...weightingOptions, buildings: e.currentTarget.value },
-              });
-            }}
-          />
-        </Group>
-      </>
-    )
-  }
-
-  function renderTransportOptions() {
-    const opts: TransportOptions = ctx.state.transportOptions;
-    const beltSelectValue = getSelectValue(opts.beltCapacity, BELT_PRESET_VALUES);
-    const pipeSelectValue = getSelectValue(opts.pipeCapacity, PIPE_PRESET_VALUES);
-
-    function dispatchBelt(value: string | null) {
-      ctx.dispatch({ type: 'UPDATE_TRANSPORT_OPTIONS', data: { ...opts, beltCapacity: value } });
-    }
-
-    function dispatchPipe(value: string | null) {
-      ctx.dispatch({ type: 'UPDATE_TRANSPORT_OPTIONS', data: { ...opts, pipeCapacity: value } });
-    }
-
-    return (
-      <>
-        <Group style={{ alignItems: 'flex-end', marginBottom: '10px' }}>
-          <Select
-            label={<LabelWithTooltip label='Belt Capacity' tooltip='Maximum items per minute any single conveyor belt can carry. When set, the solver will only produce solutions where each recipe node outputs no more than this amount per item type.' />}
-            data={BELT_TIER_OPTIONS}
-            value={beltSelectValue}
-            style={{ flex: '1 1 auto' }}
-            onChange={(value) => {
-              if (!value || value === 'disabled') {
-                dispatchBelt(null);
-              } else if (value === 'custom') {
-                dispatchBelt(opts.beltCapacity && !BELT_PRESET_VALUES.has(opts.beltCapacity) ? opts.beltCapacity : '60');
-              } else {
-                dispatchBelt(value);
-              }
-            }}
-          />
-          {beltSelectValue === 'custom' && (
-            <TextInput
-              label='Custom (items/min)'
-              className='no-spinner'
-              type='number'
-              min='1'
-              step='1'
-              value={opts.beltCapacity ?? ''}
-              style={{ flex: '1 1 auto' }}
-              onChange={(e) => dispatchBelt(e.currentTarget.value)}
-            />
-          )}
-        </Group>
-        <Group style={{ alignItems: 'flex-end' }}>
-          <Select
-            label={<LabelWithTooltip label='Pipe Capacity' tooltip='Maximum m³ per minute any single pipe can carry. When set, the solver will only produce solutions where each recipe node outputs no more than this amount per fluid type.' />}
-            data={PIPE_TIER_OPTIONS}
-            value={pipeSelectValue}
-            style={{ flex: '1 1 auto' }}
-            onChange={(value) => {
-              if (!value || value === 'disabled') {
-                dispatchPipe(null);
-              } else if (value === 'custom') {
-                dispatchPipe(opts.pipeCapacity && !PIPE_PRESET_VALUES.has(opts.pipeCapacity) ? opts.pipeCapacity : '300');
-              } else {
-                dispatchPipe(value);
-              }
-            }}
-          />
-          {pipeSelectValue === 'custom' && (
-            <TextInput
-              label='Custom (m³/min)'
-              className='no-spinner'
-              type='number'
-              min='1'
-              step='1'
-              value={opts.pipeCapacity ?? ''}
-              style={{ flex: '1 1 auto' }}
-              onChange={(e) => dispatchPipe(e.currentTarget.value)}
-            />
-          )}
-        </Group>
-      </>
-    );
   }
 
   function renderResourceInputs() {
@@ -310,26 +131,13 @@ const InputsTab = () => {
 
   return (
     <>
-      <Section>
-        <Title order={3}>Input Items</Title>
+      <CollapsibleSection title='Input Items' tooltip='Items supplied to the factory from outside. Set an amount or mark them unlimited.'>
         {renderItemInputs()}
         <Button onClick={() => { ctx.dispatch({ type: 'ADD_INPUT_ITEM' }) }}>
           + Add Input
         </Button>
-      </Section>
-      <Section>
-        <Title order={3}>Weighting Options</Title>
-        {renderWeightInputs()}
-        <Button color='red' onClick={() => { ctx.dispatch({ type: 'SET_ALL_WEIGHTS_DEFAULT', gameData: ctx.gameData }) }} style={{ marginTop: '15px' }}>
-          Reset All Weights
-        </Button>
-      </Section>
-      <Section>
-        <Title order={3}>Transport Capacity</Title>
-        {renderTransportOptions()}
-      </Section>
-      <Section>
-        <Title order={3}>Raw Resources</Title>
+      </CollapsibleSection>
+      <CollapsibleSection title='Raw Resources' tooltip='Set the available amount and weighting for each raw resource extracted from the world.'>
         <Group style={{ marginBottom: '12px' }}>
           <Button color='red' onClick={() => { ctx.dispatch({ type: 'SET_RESOURCES_TO_MAP_LIMITS', gameData: ctx.gameData }) }}>
             Set All To Maximum
@@ -345,7 +153,7 @@ const InputsTab = () => {
           style={{ marginBottom: '12px' }}
         />
         {renderResourceInputs()}
-      </Section>
+      </CollapsibleSection>
     </>
   );
 };
